pyproject.toml
[tool.poetry]
name = "auth-backend"
version = "0.1.0"
description = ""
authors = ["WATUNeed <maxconal228@gmail.com>"]
readme = "README.md"
packages = [{include = "auth_backend"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.29"
redis = "^5.0.3"
pyjwt = "^2.8.0"
alembic = "^1.13.1"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
